requirements.txt
django
djangorestframework
langchain
langchain-groq
python-dotenv
langchain-community
djangorestframework-simplejwt
django-cors-headers
openai
langchain-core